requirement.txt
aiofiles==23.2.1
albumentations==1.4.2
annotated-types==0.7.0
anyio==4.4.0
beautifulsoup4==4.12.3
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
einops==0.2.0
fastapi==0.112.0
ffmpy==0.4.0
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
gdown==4.4.0
gradio==4.40.0
gradio_client==1.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
imageio==2.34.2
imgaug==0.4.0
importlib_resources==6.4.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
lmdb==1.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pdf2image==1.17.0
pillow==10.2.0
prefetch-generator==1.0.1
pybind11==2.13.1
pyclipper==1.3.0.post5
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
reportlab==4.2.2
requests==2.32.3
rich==13.7.1
ruff==0.5.6
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
semantic-version==2.10.0
shapely==2.0.5
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
sympy==1.12
threadpoolctl==3.5.0
tifffile==2024.7.24
tomlkit==0.12.0
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.5
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.5
vietocr==0.3.13
websockets==12.0
